vcf_transform/requirements.txt
python=3.9.4
pyvcf=0.6.8